Updated Emotions_enums page
