+++ mobile/voice_journal_app/lib/Emotions_enums.dart
@@ -1,0 +1,8 @@
+enum Emotions {
+ 	happiness, // id = 0
+  sadness, // id = 1
+  anger, // id = 2
+  fear, // id = 3
+  disgust, // id = 4
+  surprise, // id = 5
+}
